Mix Shadows button added
Mix shadow functionality moved to check list UI button
