+++ Assets/Scripts/UI/MixShadowButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 463b8717c8aa05a4eb4e22b5f30cb4b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
